Added views, fixed obvious errors.
The JSON feed from MLB only provides the next game, not an empty
json array when the next game is in the future, so there's a check that
the first game in the list is actually for today.

My RedisCache implementation feels a little hacky and too large for
what is effectively a wrapper around a cache. It works, but I'd like to
look at how redis-rb stores a binary type so we don't need to coerce a
string to a boolean.